Add the ability to edit the hero name
Users should be able to edit the hero name in an <input> textbox.  The
textbox should both display the hero's `name` property and update that
property as the user types.

We use a two-way binding between the <input> form element and the
`hero.name` property.

See https://angular.io/tutorial/toh-pt1#edit-the-hero-name for more
details.