Display a list of heroes
In this commit we add a public property in `AppComponent` that exposes
the heroes for binding.

    heroes = HEROES;

The `heroes` type isn't defined because TypeScript infers it from the
`HEROES` array.

Then we use the built-in directive `*ngFor` to iterate over the
component's `heroes` array and renders an instance of the template for
each hero in that array.

The `*` prefix to `ngFor` is a critical part of this syntax.  It
indicates that the `<li>` element and its children constitute a master
template.

See https://angular.io/tutorial/toh-pt2 for more details.