Add the ability to select a hero and display its details
The app displays a list of heroes as well as a single hero in the
details view.  But the list and the details view are not connected.
When users select a hero from the list, the selected hero should appear
in the details view.  This UI pattern is known as "master/detail."  In
this case, the _master_ is the heroes list and the _detail_ is the
selected hero.

In this commit, we connect the master to the detail through a
`selectedHero` component property, which is bound to a click event.

To learn more about event binding, visit the User Input page at
https://angular.io/guide/user-input and the Event binding section of the
Template Syntax page at
https://angular.io/guide/template-syntax#event-binding

Read more about `ngIf` and `ngFor` in the Structural Directives page at
https://angular.io/guide/structural-directives and the Built-in
directives section of the Template Syntax page at
https://angular.io/guide/template-syntax#directives.

https://angular.io/tutorial/toh-pt2#selecting-a-hero