Use multiple components
The `AppComponent` is doing everything at the moment.  In the beginning,
it showed details of a single hero.  Then it became a master/detail form
with both a list of heroes and the hero detail.  Soon there will be new
requirements and capabilities.  We can't keep piling features on top of
features in one component; that's not maintainable.

We'll need to break it up into sub-components, each focused on a
specific task or workflow.  Eventually, the `AppComponent` could become
a simple shell that hosts those sub-components.

https://angular.io/tutorial/toh-pt3