Add a hero service
As the Tour of Heroes app evolves, we'll add more components that need
access to hero data.

Instead of copying and pasting the same code over and over, we'll create
a single reusable data service and inject it into the components that
need it.  Using a separate service keeps components lean and focused on
supporting the view, and makes it easy to unit-test components with a
mock service.

Because data services are invariably asynchronous, we'll finish the page
with a Promise-based version of the data service.

https://angular.io/tutorial/toh-pt4