Split the AppComponent
The current app loads `AppComponent` and immediately displays the list
of heroes.

The revised app should present a shell with a choice of views (Dashboard
and Heroes) and then default to one of them.

The `AppComponent` should only handle navigation, so you'll move the
display of heroes out of `AppComponent` and into its own
`HeroesComponent`.

https://angular.io/tutorial/toh-pt5#splitting-the-appcomponent